requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pydantic==1.7.3
python-multipart==0.0.5
PyYAML==5.3.1
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3

dictdiffer~=0.8.1